tdd_fastapi/requirements.txt
fastapi
uvicorn
asyncpg==0.22.0
tortoise-orm==0.17.3
